+++ src/internal/services/project_manager/config.go
@@ -1,0 +1,85 @@
+package service
+
+import (
+	rest_api_conf "sm-box/internal/services/project_manager/transports/rest_api/config"
+	"sm-box/pkg/core/components/configurator"
+	"sm-box/pkg/core/components/tracer"
+)
+
+// Config - конфигурация сервиса.
+type Config struct{}
+
+// ConfigTransports - конфигурация транспортной части сервиса
+type ConfigTransports struct {
+	RestAPI *rest_api_conf.Config `json:"rest_api" yaml:"RestAPI" xml:"RestAPI"`
+}
+
+// Read - чтение конфигурации.
+func (conf *Config) Read() (err error) {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.Error(err).FunctionCallFinished() }()
+	}
+
+	var (
+		c       configurator.Configurator[*Config]
+		profile = configurator.PrivateProfile{
+			Dir:      "/",
+			Filename: "service.xml",
+		}
+	)
+
+	if c, err = configurator.New[*Config](conf); err != nil {
+		return
+	} else if err = c.Private().Profile(profile).Init(); err != nil {
+		return
+	}
+
+	if err = conf.FillEmptyFields().Validate(); err != nil {
+		return
+	}
+
+	return
+}
+
+// FillEmptyFields - заполнение пустых полей конфигурации
+func (conf *Config) FillEmptyFields() *Config {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.FunctionCallFinished(conf) }()
+	}
+
+	return conf
+}
+
+// Default - запись стандартной конфигурации.
+func (conf *Config) Default() *Config {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.FunctionCallFinished(conf) }()
+	}
+
+	return conf
+}
+
+// Validate - валидация конфигурации.
+func (conf *Config) Validate() (err error) {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.Error(err).FunctionCallFinished() }()
+	}
+
+	return
+}